모임가입해야 todo이용가능

+++ src/main/java/us/flower/dayary/domain/ToDoWrite.java
@@ -55,6 +55,6 @@ public class ToDoWrite extends DateAudit{
 	
 	private String count;
 	
-	private int progress;
+	private double progress;
 	
 }
